requirements.txt
sanic
Sanic-Cors
python-socketio
sanic-jwt
gunicorn
pytest-sanic
aiocassandra